done the impl

+++ swigatto/src/main/java/com/swigatto/swigatto/model/Cart.java
@@ -19,7 +19,7 @@ public class Cart {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     int id;
 
-    int cartTotal;
+    double cartTotal;
 
     @OneToOne
     @JoinColumn
