+++ swigatto/src/main/java/com/swigatto/swigatto/model/FoodItem.java
@@ -18,7 +18,7 @@ public class FoodItem {
 
     int requiredQuantity;
 
-    int totalCost;
+    double totalCost;
 
     @ManyToOne
     @JoinColumn
